requirements.txt
pillow
numpy
RPi.GPIO
spidev
psutil
requests
feedparser
flask
openai
